This backports the REXML 1.9.3 patch to REE for CVE-2014-8090
Patchfile is based on this commit to 1.9.3:
https://github.com/ruby/ruby/commit/e70ce4c096e3d1557a9bb5b7e89a9bb613b0794c

I selectively diffed the patched 1.9.3 lib/rexml/document.rb and
test/rexml/test_document.rb to only select the changes needed as suggested
by the monkey patch at
https://www.ruby-lang.org/en/news/2014/11/13/rexml-dos-cve-2014-8090/

thus, some tests that were added to 1.9.3 for this CVE weren't added
as REE 1.8.7 didn't have a test for XML_WITH_NESTED_PARAMETER_ENTITY
to begin with.

I think this should work. But, I haven't built it yet.